implement MVC Pattern by moving all data manipulation inside struct in model.swift file.
